warning from converting DWORD bitsPerSample to native WORD storage
In the waveout structure specifications, bits per sample is stored as a 16-bit integer, so maybe the function should be reading from a WORD type rather than DWORD type.  Besides, it doens't look like the waveOut code is used much anymore, so in the worst case, this just fixes a warning.