Swap alternating lines -- 10325476 instead of 01234567.
I know this may look uglier like I'm hijacking the endian or something, but there is a reason I'm uglifying the order.  I notice the assignments all start from the inp2[] vector then read over to inp1[] once the array index to inp2[] flows under, and with the more big-endian 01234567 order of the multiply-adds we were saying inp2 then inp1 then inp2 again...this way the contiguity of the arrays when memory access leaks over to the alternating vector can be seen more easily.