Please avoid changing the plugin spec definitions in that manner.
Changing the plugin spec headers by itself can be good for portability but not when it contradicts the definition of the plugin spec functions and things like that--this makes everybody have to update either their correct headers or their own source code, if not both, to test Azimer's code or things like that in their own audio plugins and really impedes spec compatibility with these macro functions.

Look at the bottom of this commit to see my solution for Xbox:  Give EXPORT and CALL null macro definitions.