Use halfword-endian-swap macro from my_types.h for spec-fixed endian.
RDRAM and other things shared between the plugin and the host interface emulator are in the host-defined endian--which for emulators like Project64 is usually little-endian 32 bits.  The endian adjustment macros I had in "my_types.h" address this.

I do not think the ^ 2 is readable because it's not important to N64 info/emulation...just a requirement to port something to zilmar's spec until we get a big-endian emulator.