Remove the fourth parameter--merge it in to the second as the pointer.
I always wanted to do this from the start but was unable due to all the XOR endian swapping.  (The starting address of an off-setted vector might need to go backwards to previous elements, which is not something that XOR lets you do.)  Now that the endian stuff is taken care of, we can do this.