Avoid the Windows API type `BYTE' for non-Windows-relevant code.
While it could be inferred that "BYTE" be a generic type for unsigned chars on any operating system--either Microsoft Windows or any other--the type name itself still has traditional problems.

First, it is a reserved type enforced within the <windows.h> set of headers.  Because of this, defining your own `BYTE` type on other operating systems runs the risk of conflicting with standard, potentially built-in types on other compilers.

Second, it is not standard practice in C to use all uppercase letters for type names.

Third, nearly anyone who is familiar with `BYTE` as a standard type knows that it is necessary for Windows API/OS functions, such as memory management on the Windows operating system.  The usage of this type gives a false impression that such things are relevant.

Fourth, `char` is not always a byte.  The C standard requires neither that `char` be 8 bits, nor that the implementation equates the size of a char to that of a byte on the machine.

Fifth, `unsigned char` is a standard type that is more readable than `BYTE` or `byte` here, because it is the smallest unit of memory that C will let us address.

Sixth, `BYTE` and `byte` never specify our needs for the size of a byte.  More specific is `u8` or `uint8_t`.

Seventh, whether `BYTE` or `byte` typedefs are pertinent to the target system being emulated or the environment system used to emulate is left open to interpretation definitively.  The only non-biased, portable intention is that it refers to `BYTE` as in the size of a byte on the MIPS architecture, but that requires custom type definitions to guarantee that this holds true, such as a type named MIPS_byte which, again, makes usage of the Windows-pertinent type names such as "BYTE" obsolete for that purpose.