Debatably, a macro or some constant here can be of benefit.
Of questionable value, but sometimes people will see 0x20 or 32 and really not see the relevance right away behind the value (in this case the total number of commands).  A constant such as a macro can sometimes be more readable; if it isn't, then it still promotes ease of adjustment for things like debugging issues.