Added base structure for common buffer emulation
Added the base structure to allow for an API agnostic handling of the AI
buffer emulation.  This commit is basically a bridge between the old
methodology and the new so any other PRs won't interfere as much if they
modify the same code.  Some regression is possible but won't be fixed.
Updated WIP to 6 to indicate the significance of the coming change.